Add stock_health under data/
