+++ stock_market/data/__init__.py
@@ -3,7 +3,7 @@ from pathlib import Path as _Path
 import pandas as _pandas
 
 from stock_market.data._ipo import IPO
-from stock_market.data._stocks import get_ticker
+from stock_market.data._stocks import get_ticker, stock_health
 from stock_market.data.reddit.trends import get_reddit_top_posts
 
 # S&P data
